+++ src/index.ts
@@ -1,3 +1,4 @@
-export { parse, parseGames, parseGame } from './pgn-parser'
+export { parse, parseGames, parseGame, SyntaxError } from './pgn-parser'
+// export { parse, parseGames, parseGame, SyntaxError } from './pgn-parser'
 export { split, SplitGame } from './split-games'
 export * from './types'