+++ src/types.ts
@@ -31,7 +31,6 @@ type MessagesObject = { messages: Message[] }
 
 export type GameComment = { comment?: string, colorArrows?: string[], colorFields?: string[],
     clk?: string, egt?: string, emt?: string, mct?: string, eval?: string }
-export type ParseTreeOrArray = ParseTree | ParseTree[]
 export type PgnMove = {
     drawOffer: boolean;
     moveNumber: number,